Rework express setup to be declarative and synchronous.
Moves the async setup and listen back to server.js
